backend/requirements.txt
# FastAPI และ dependencies หลัก
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3

# CORS support
python-dotenv==1.0.1

# HTTP client สำหรับเรียก LLM API
httpx==0.28.1
